requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.19
pytest==7.4.0
mariadb==1.1.7
python-jose[cryptography]==3.3.0
